Graph based clustering using python
This code is performing spectral clustering on a dataset of Dow Jones index data, in order to group the data into clusters. The data is first cleaned and preprocessed, then a Principal Component Analysis (PCA) is applied to reduce the dimensionality of the data. The Calinski Harabasz index is then used to evaluate the quality of clustering for a range of possible number of clusters, and the number of clusters that results in the highest CH index is chosen. Spectral clustering is then applied to the data using two different affinity functions, radial basis function (RBF) and nearest neighbors, and the resulting clusters are visualized. The final visualizations show the data points in the two-dimensional feature space, colored according to their cluster membership.